+++ alphastats/gui/utils/import_helper.py
@@ -9,7 +9,7 @@ from streamlit.runtime.uploaded_file_manager import UploadedFile
 
 from alphastats.DataSet import DataSet
 from alphastats.gui.utils.options import SOFTWARE_OPTIONS
-from alphastats.loader.MaxQuantLoader import BaseLoader, MaxQuantLoader
+from alphastats.loader.maxquant_loader import BaseLoader, MaxQuantLoader
 
 
 def load_proteomics_data(uploaded_file, intensity_column, index_column, software):
