+++ alphastats/loader/diann_loader.py
@@ -1,4 +1,4 @@
-from alphastats.loader.BaseLoader import BaseLoader
+from alphastats.loader.base_loader import BaseLoader
 
 
 class DIANNLoader(BaseLoader):
