+++ alphastats/loader/maxquant_loader.py
@@ -3,7 +3,7 @@ from typing import Union
 import numpy as np
 import pandas as pd
 
-from alphastats.loader.BaseLoader import BaseLoader
+from alphastats.loader.base_loader import BaseLoader
 
 
 class MaxQuantLoader(BaseLoader):
