+++ alphastats/loader/spectronaut_loader.py
@@ -5,7 +5,7 @@ from typing import Union
 import numpy as np
 import pandas as pd
 
-from alphastats.loader.BaseLoader import BaseLoader
+from alphastats.loader.base_loader import BaseLoader
 
 SPECTRONAUT_COLUMN_DELIM = "."
